loader: Match only if sys._MEIPATH is a prefix of package-path.
This fixes loading of plugin-type modules at run-time of the frozen
application for this case:
If the plugin path is one character longer than sys._MEIPATH
(e.g. "$PWD/p/plugin_1" and "$PWD/dist/main"),
the plugin relative-imports a sub-module (of the plugin)
and the frozen application contains a module of the same name,
the frozen application module was imported.

This was caused by the FrozenImporter not checking at all whether the
package's __path__ element is a prefix of sys._MEIPATH.

Co-authored-by: Hartmut Goebel <h.goebel@crazy-compilers.com>
Fix #4141